waybar: Show window title for the current output
This is a hack since I'm forced to display the same bar on my laptop
screen and my external monitor when I really want the laptop bar to go
away when plugged into an external monitor.

In that case, the window title changing can be seen in my peripheral
vision and is very distracting.